Move cli specific methods from N5Helpers to PainteraCommandLineArgs
